GUI: Move transactions filter to own class
Also clean up some of the logic to reduce re-rendering
